refactor: useTextFormat context -> hook
